fastify.listen: Return promise if callback is not provided
fastify.listen handled errors inconsistently when no callback was
provided.  This was partly due to the fact that we cannot test uncaught
exceptions from tap, and the exception occurs in a callback.  To deal
with this we return a Promise from fastify.listen when no callback is
provided.  Also added full test coverage for running fastify.listen
without a callback.

This has is a breaking change to error handling when no callback is
provided.  In the past when fastify.listen errored without a callback it
would throw an exception, likely leading to node.js exiting.  Now the
returned promise will reject, if nothing catches the error a warning
will be printed.

Fixes #501