Improve handling undefined payloads
According to RFC 7230 Section 3.3.2, there are cases when a server must
not send a `Content-Length` header (such as with a 204 response). This
change makes Fastify follow the spec more closely by passing the payload
through the `onSend` hooks as-is and not setting the `Content-Length`
header on a response with an `undefined` payload.